+++ src/main/java/no/uio/ifi/trackfind/backend/data/providers/gdc/GDCPage.java
@@ -23,7 +23,7 @@ public class GDCPage implements Page {
      * {@inheritDoc}
      */
     @Override
-    public long getPagesTotal() {
+    public int getPagesTotal() {
         return getSubpage().getPagination().getPages();
     }
