+++ src/main/java/no/uio/ifi/trackfind/backend/data/providers/icgc/ICGCPage.java
@@ -25,7 +25,7 @@ public class ICGCPage implements Page {
      * {@inheritDoc}
      */
     @Override
-    public long getPagesTotal() {
+    public int getPagesTotal() {
         return getPagination().getPages();
     }
